+++ train.py
@@ -79,7 +79,7 @@ def validate(model, data_loader):
             label = data[2].to(device)
             scale = data[3].to(device)
 
-            with amp.autocast("cuda", enabled=use_amp, dtype=torch.float16):
+            with amp.autocast("cuda", enabled=use_amp, dtype=torch.bfloat16):
                 logits = model(input_img, cropped_img, scale)
             y_pred.extend(logits.sigmoid().flatten().tolist())
             y_true.extend(label.flatten().tolist())
@@ -144,7 +144,7 @@ if __name__ == '__main__':
 
             model.set_input(data)
             model.optimizer.zero_grad(set_to_none=True)
-            with amp.autocast("cuda", enabled=use_amp, dtype=torch.float16):
+            with amp.autocast("cuda", enabled=use_amp, dtype=torch.bfloat16):
                 model.forward()
                 loss = model.get_loss()
